src/requirements.txt
pandas==1.4.1
python-decouple==3.6
requests==2.27.1